requirements.txt
aiohttp>=3.11.11
aiosignal>=1.3.1
async-timeout>=5.0.1
attrs>=24.3.0
certifi>=2024.12.14
chardet>=5.2.0
charset-normalizer>=3.4.1
discord.py>=2.4.0
frozenlist>=1.5.0
idna>=3.10
lenny>=0.1.3
multidict>=6.1.0
typing_extensions>=4.12.2
urllib3>=2.3.0
websockets>=14.1
yarl>=1.18.3
